pyproject.toml
[build-system]
requires = [
  "setuptools>=62.1",
  "setuptools_scm[toml]>=8.0.0",
  "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "kcexo"
description = "Collection of randome exoplanet observation planning and analysis scripts and libraries"
requires-python = ">=3.12"
readme = { file = "README.rst", content-type = "text/x-rst" }
license = { file = "licenses/LICENSE.rst" }
authors = [
  { name = "Daniel Kustrin", email = "daniel_kustrin@yahoo.co.uk" },
]
dependencies = [
  "numpy",
  "scipy",
  "pyvo",
  "matplotlib",
  "astropy",
  "astroplan",
  "astroquery",
  "wxpython",
  "pypubsub",
  "jsonschema"
]
dynamic = ["version"]

[project.scripts]
kc_comp_stars = "kcexo.ui.comp_stars.exo_comp_stars:main"
kc_planner = "kcexo.ui.planner.exo_planner:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-doctestplus",
    "pytest-cov",
    "pytest-datadir"
]
docs = [
    "sphinx",
    "sphinx-automodapi",
]

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
include = ["kcexo*"]
exclude = ["kcexo._dev*", "kcexo.*.play*"]

[tool.setuptools_scm]
version_file = "kcexo/_version.py"

[tool.pytest.ini_options]
testpaths = [
  "kcexo",
  "docs",
]
doctest_plus = "enabled"
text_file_format = "rst"
addopts = "--doctest-rst"

[tool.coverage.run]
omit = [
 "kcexo/__init*",
 "kcexo/conftest.py",
 "kcexo/*setup_package*",
 "kcexo/tests/*",
 "kcexo/*/tests/*",
 "kcexo/extern/*",
 "kcexo/version*",
 "*/kcexo/__init*",
 "*/kcexo/conftest.py",
 "*/kcexo/*setup_package*",
 "*/kcexo/tests/*",
 "*/kcexo/*/tests/*",
 "*/kcexo/extern/*",
 "*/kcexo/version*",
]

[tool.coverage.report]
exclude_lines = [
  # Have to re-enable the standard pragma
  "pragma: no cover",
  # Don't complain about packages we have installed
  "except ImportError",
  # Don't complain if tests don't hit assertions
  "raise AssertionError",
  "raise NotImplementedError",
  # Don't complain about script hooks
  "def main(.*):",
  # Ignore branches that don't pertain to this version of Python
  "pragma: py{ignore_python_version}",
  # Don't complain about IPython completion helper
  "def _ipython_key_completions_",
  # typing.TYPE_CHECKING is False at runtime
  "if TYPE_CHECKING:",
  # Ignore typing overloads
  "@overload",
]
